requirements.txt
# Core
django==4.2.9

# UI & Admin UI
django-jazzmin==3.0.1
django-admin-black==1.0.19

# Tools
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
django-extensions==3.2.3
requests==2.32.3
pandas==2.2.3
graphviz==0.20.3
astor==0.8.1
pre_commit==4.2.0
djangorestframework-camel-case==1.4.2
django-filter==25.1
model-bakery==1.20.5
django-tables2==2.7.5
crispy-bootstrap5==2025.6
django-crispy-forms==2.4

# API docs via Swagger/OpenAPI
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.9.1
drf-yasg==1.21.10

# Deployment
whitenoise==6.7.0
gunicorn==23.0.0

# DB
#psycopg2-binary==2.9.9
#mysqlclient==2.1.1
django-dbbackup==4.2.1

# Linters, pycodestyles
ruff==0.13.2
